@@ -6,55 +6,64 @@
     "minimum-stability": "dev",
     "prefer-stable": false,
     "require": {
-        "php": ">=8.0",
+        "php": ">=8.1",
         "ext-ctype": "*",
         "ext-iconv": "*",
         "ext-intl": "*",
         "ext-pgsql": "*",
-        "doctrine/doctrine-bundle": "^2.4",
-        "doctrine/doctrine-migrations-bundle": "^3.1",
-        "doctrine/orm": "^3.0",
-        "sensio/framework-extra-bundle": "^6.2",
-        "symfony/console": "^6.0",
-        "symfony/dotenv": "^6.0",
+        "doctrine/doctrine-bundle": "^2.7",
+        "doctrine/doctrine-migrations-bundle": "^3.2",
+        "doctrine/orm": "^2.13",
+        "sensio/framework-extra-bundle": "^6.1",
+        "symfony/console": "^6.1",
+        "symfony/dotenv": "^6.1",
         "symfony/flex": "^1.3.1",
         "symfony/framework-bundle": "^6.0",
         "symfony/monolog-bundle": "^3.7",
-        "symfony/proxy-manager-bridge": "^6.0",
-        "symfony/runtime": "^6.0",
-        "symfony/serializer": "^6.0",
-        "symfony/twig-bundle": "^6.0",
+        "symfony/property-access": "^6.1",
+        "symfony/proxy-manager-bridge": "^6.1",
+        "symfony/runtime": "^6.1",
+        "symfony/serializer": "^6.1",
+        "symfony/twig-bundle": "^6.1",
         "symfony/webpack-encore-bundle": "^1.13",
-        "symfony/yaml": "^6.0",
+        "symfony/yaml": "^6.1",
         "twig/extra-bundle": "^2.12|^3.0",
         "twig/twig": "^2.12|^3.0"
     },
     "require-dev": {
         "ext-sqlite3": "*",
         "phpunit/phpunit": "10.0.x-dev",
-        "symfony/browser-kit": "^6.0",
-        "symfony/css-selector": "^6.0",
-        "symfony/debug-bundle": "^6.0",
+        "symfony/browser-kit": "^6.1",
+        "symfony/css-selector": "^6.1",
+        "symfony/debug-bundle": "^6.1",
         "symfony/maker-bundle": "dev-main",
-        "symfony/phpunit-bridge": "6.0.x-dev",
-        "symfony/stopwatch": "^6.0",
-        "symfony/web-profiler-bundle": "^6.0"
+        "symfony/phpunit-bridge": "6.1.x-dev",
+        "symfony/stopwatch": "^6.1",
+        "symfony/web-profiler-bundle": "^6.1"
     },
     "config": {
         "optimize-autoloader": true,
+        "platform": {
+            "php": "8.1"
+        },
         "preferred-install": {
             "*": "dist"
         },
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "symfony/flex": true,
+            "symfony/runtime": true
+        }
     },
     "autoload": {
         "psr-4": {
-            "App\\": "src/"
+            "App\\": "src/App",
+            "ArtefactGuide\\": "src/ArtefactGuide"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "App\\Tests\\": "tests/"
+            "Tests\\": "tests/"
         }
     },
     "replace": {
@@ -80,7 +89,7 @@
     "extra": {
         "symfony": {
             "allow-contrib": false,
-            "require": "^6.0"
+            "require": "^6.1"
         }
     }
 }
